// src/gs/interfaces/cols.interface.ts
export interface Cols {
    readonly COLAlloc: string;
    readonly Quantity: number;
    readonly MinQuantity: number;
    readonly MaxQuantity:number;
    readonly RoutingId: string;
    readonly ProductId: string;
    readonly ProductName: string;
    readonly LatestDesiredDeliveryDate: number;
    readonly ProductSpecificationId: string;
    readonly ResourceGroupIds: string;
    readonly ImgPlannedStatus: string;
}